+++ src/test/java/com/rakovets/course/java/core/practice/oop/principles/battle_ground/WarriorTest.java
@@ -1,4 +1,4 @@
-package com.rakovets.course.java.core.practice.oop.principles.battle_groundTest;
+package com.rakovets.course.java.core.practice.oop.principles.battle_ground;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
@@ -8,7 +8,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.util.stream.Stream;
 
-class ArcherTest {
+class WarriorTest {
     static Enemy enemy = new Enemy(100) {
         @Override
         public void attackEnemy(Hero hero) {
@@ -18,29 +18,29 @@ class ArcherTest {
 
     static Stream<Arguments> attackEnemyTest() {
         return Stream.of(
-                Arguments.of(85, enemy));
+                Arguments.of(95, enemy));
     }
 
     @ParameterizedTest
     @MethodSource("attackEnemyTest")
     void attackEnemyTest(int expected, Enemy enemy) {
-        enemy.setHealth(100);
-        Archer archer = new Archer("Archer", 100);
-        archer.attackEnemy(enemy);
+        Warrior warrior = new Warrior("Warrior", 50);
+        warrior.attackEnemy(enemy);
         assertEquals(expected, enemy.getHealth());
     }
 
-    static Stream<Arguments> rainOfArrowsTest() {
+    static Stream<Arguments> doomPunchTest() {
         return Stream.of(
-                Arguments.of(55, enemy));
+                Arguments.of(65, enemy));
     }
 
     @ParameterizedTest
-    @MethodSource("rainOfArrowsTest")
-    void rainOfArrowsTest(int expected, Enemy enemy) {
-        Archer archer = new Archer("Archer", 100);
+    @MethodSource("doomPunchTest")
+    void doomPunchTest(int expected, Enemy enemy) {
+        Warrior warrior = new Warrior("Warrior", 50);
         enemy.setHealth(100);
-        archer.rainOfArrows(enemy);
+        warrior.doomPunch(enemy);
         assertEquals(expected, enemy.getHealth());
     }
+
 }
